"ads_2_remedial_last: 0/6 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ assignments/ADS - 2 Remedial/Solution.java
@@ -11,10 +11,10 @@ class Solution {
         EdgeWeightedGraph ewgraph = new EdgeWeightedGraph(circles);
         for (int i = 0; i < routes; i++) {
             String[] route = scan.nextLine().split(" ");
-            Edge edg = new Edge(Integer.parseInt(route[0]),Integer.parseInt(route[1]),Double.parseDouble(route[2]));
+            Edge edg = new Edge(Integer.parseInt(route[0]),Integer.parseInt(route[1]),Integer.parseInt(route[2]));
             ewgraph.addEdge(edg);
         }
         KruskalMST krush = new KruskalMST(ewgraph);
-        System.out.format("%.5f", krush.weight());
+        System.out.format("%d", krush.weight());
     }
 }
